Toast added

+++ src/Components/MySubToy.jsx
@@ -1,18 +1,42 @@
 import { BsStarFill } from "react-icons/bs";
 import { RiDeleteBin6Fill } from "react-icons/ri";
+import Swal from "sweetalert2";
 
-const MySubToy = ({main}) => {
+const MySubToy = ({main, setData, datas}) => {
     const {_id, category, img, price, rating,  } = main;
    const handleDelete = (id) =>{
-      console.log(id);
+    Swal.fire({
+        title: 'Are you sure?',
+        text: "You won't be able to revert this!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Yes, delete it!'
+      }).then((result) => {
+        if (result.isConfirmed) {
+         
+        fetch(`https://assign11-server.vercel.app/toy/${id}`,{
+            method:"DELETE"
+          })
+          .then(res => res.json())
+          .then( data =>{
+            console.log(data)
+            if(data.deletedCount > 0){
+                Swal.fire(
+                    'Deleted!',
+                    'Your file has been deleted.',
+                    'success'
+                  )
+              const remaining = datas.filter(m=>m._id !== _id);
+              setData(remaining)
+            }
+          })
+        }
+      })
 
-      fetch(`https://assign11-server.vercel.app/toy/${id}`,{
-        method:"DELETE"
-      })
-      .then(res => res.json())
-      .then(data =>{
-        console.log(data)
-      })
+      
+
    }
     return (
 
@@ -26,10 +50,11 @@ const MySubToy = ({main}) => {
            
            
             <button className="btn btn-warning">Edit</button>
-            <button onClick={()=>handleDelete(_id)} className="btn bg-white hover:bg-white"><RiDeleteBin6Fill  className="text-black  text-2xl"/> </button>
+            <button onClick={ ()=>handleDelete(_id) } className="btn bg-white hover:bg-white"><RiDeleteBin6Fill  className="text-black  text-2xl"/> </button>
         </div>
         </div>
     );
 };
 
 export default MySubToy;
+
